User entity  mdoification

+++ src/main/java/com/ImperioElevator/ordermanagement/dao/OrderDao.java
@@ -26,5 +26,5 @@ public interface OrderDao extends Dao<Order>{
 
     Order getOrderWithExtraProducts(Long orderId);
 
-    Long deleteUnconfirmedOrderByEmail()throws SQLException;
+    Integer deleteUnconfirmedOrderByEmail()throws SQLException;
 }
